separate AbstractPotential
For runtime type checking
