+++ src/galdynamix/potential/_potential/__init__.py
@@ -1,14 +1,16 @@
 """galdynamix: Galactic Dynamix in Jax."""
 
 
-from . import base, builtin, composite, param
+from . import base, builtin, composite, core, param
 from .base import *
 from .builtin import *
 from .composite import *
+from .core import *
 from .param import *
 
 __all__: list[str] = []
 __all__ += base.__all__
+__all__ += core.__all__
 __all__ += composite.__all__
 __all__ += param.__all__
 __all__ += builtin.__all__
